requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.2
Django==4.1.6
django-appconf==1.0.5
django-bootstrap4==22.3
django-compressor==4.3.1
django-countries==7.5.1
django-crispy-forms==2.0
django-jazzmin==2.6.0
django-libsass==0.9
django-phonenumber-field==7.0.2
install==1.3.5
libsass==0.22.0
rcssmin==1.1.1
rjsmin==1.2.1
soupsieve==2.4
sqlparse==0.4.3
typing_extensions==4.4.0
whitenoise==6.3.0
